[KNOWAGE-5166] fixing issue on line grouping

+++ knowagecockpitengine/src/main/java/it/eng/spagobi/engine/chart/util/DataSetTransformer.java
@@ -852,8 +852,8 @@ public class DataSetTransformer {
 		if (drillOrder != null) {
 			for (String key : drillOrder.keySet()) {
 				Map<String, String> keyMapper = (Map<String, String>) drillOrder.get(key);
-				if (!keyMapper.get("orderColumn").equals("") && !keyMapper.get("orderColumn").equals(categorieColumns.get("column"))
-						&& !drillOrder.containsKey(keyMapper.get("orderColumn"))) {
+				if (keyMapper.get("orderColumn") != null && !keyMapper.get("orderColumn").equals("")
+						&& !keyMapper.get("orderColumn").equals(categorieColumns.get("column")) && !drillOrder.containsKey(keyMapper.get("orderColumn"))) {
 					dataColumnsMapper.remove(keyMapper.get("orderColumn").toLowerCase());
 				}
 			}
